% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_proc_LP.R
\name{NW.LongRunVariance}
\alias{NW.LongRunVariance}
\title{Newey–West Long-Run Variance Estimator}
\usage{
NW.LongRunVariance(X, q)
}
\arguments{
\item{X}{A vector or a matrix representing the time series.
If \code{X} is a matrix of size \eqn{T \times k}, the function returns a
\eqn{k \times k} long-run covariance matrix.}

\item{q}{A non-negative integer specifying the Newey–West truncation lag.
If \code{q = 0}, the function returns the contemporaneous covariance
\code{autocov(X, 0)}.}
}
\value{
A scalar (if \code{X} is univariate) or a matrix (if \code{X} is
multivariate) giving the Newey–West long-run variance or covariance estimate.
}
\description{
Computes the Newey–West estimator of the long-run variance
(or long-run covariance matrix) of a univariate or
multivariate time series.
}
\examples{
# Univariate example
set.seed(123)
x <- rnorm(200)
NW.LongRunVariance(x, q = 4)

# Multivariate example
X <- cbind(rnorm(200), rnorm(200))
NW.LongRunVariance(X, q = 5)

}
